modified - migrations updated - cart, added remove item

+++ Udemy.Core/Entities/Section.cs
@@ -8,7 +8,7 @@ namespace Udemy.Core.Entities
         [StringLength(20)]
         public string Title { get; set; }
         //public int Order { get; set; } 
-        public int Duration { get; set; }
+        public decimal Duration { get; set; }
         public int NoLessons { get; set; }
 
         [ForeignKey("Course")]
